[update] repository 폴더 경로 정리
